Add firmware update failure handling
If the firmware update file is accepted by
the Ewon device, the thread is blocked and
none of the code after that is executed. It
is safe to assume that if the connector
continues to execute after a successful
firmware file download, the firmware file
was not accepted by the Ewon and is invalid.