Added deploy to s3 action code
This commit adds a "deploy" application and a github action to upload
production artifacts to S3; to deploy artifacts to s3 based staging
area. The application is a separate Node.js application that utilizes
both the github REST API and S3 REST API. The application is specific
to the Cumulocity Connector because of the production artifacts are
unique to this solution.