Modify custom certificate support
Modify custom certificate support such that it
now is an opt-in/opt-out feature via the config
file.

If custom URL certificate support is not enabled,
the text contents of the default Cumulocity Root
CA will be written to a file and used. If custom
URL certificate support is enabled, the custom
certificate download URL from the config file
will be downloaded to a file and used.

Existing users will have their configuration
automatically migrated from the former field
`CertificateUrl` to the new `CustomCertificateUrl`
field with the `CustomCertificateUrlEnabled`
value set to true.