Update pom.xml
Updates the pom.xml file to bring it in line with
the latest available version from the upstream starter
project. This excludes the changes to project assembly,
which are to be implemented at a later date along with
corresponding other workflow/structure changes. The
Checkstyle validation has also been excluded at this time
due to ongoing discussion and refinement regarding the
strictness of certain checks.

Notably, this downgrades the JUnit dependency to version
3.8.1. While it is preferred to use newer versions under
most circumstances, there is no reason for this project
to deviate from the upstream project. There are (currently)
no test implementations in this project, and there are
differences between the JUnit major versions that would
result in tests being incompatible between versions.