Add MQTT manager last known status
Added a mechanism for storing and retrieving a last
known MQTT status code from the manager class.

This allows the manager to inject its own logic as
well, such as considering the MQTT connection status
to be unknown if the connector detects the network
is down before an MQTT status change occurs.