Update CConnectorDataMgr poll logic
Updated the polling logic in CConnectorDataMgr.java
so that the MQTT manager's last known status code is
checked. If the MQTT manager does not indicate that
it is connected, data polling will be skipped.

This prevents the build-up of data in the connector's
memory which could be lost should the Ewon lose power
or otherwise reboot before the network condition has
resolved.