Add config data processing mode support
Added support for configuring the data processing mode
specified to Cumulocity for data payloads (standard or
aggregated JSON). This required the differentiation
between types of payloads being sent, thus the reason
for the previously added data processing mode enum-like
class.

Because the sendMessageWithChildDeviceRouting method
was being adapted to require a message type input, the
JSON-specific method was also combined to reduce the
amount of unnecessary/duplicate code present.